Correct a field type in the locus model and add the top x gene level abundance list in sample view and scatter plot for sum jpm circrna vs canonical junctions

+++ core/models/locus.py
@@ -14,7 +14,9 @@ class Locus(models.Model):
     browser_string = models.TextField(
         blank=True
     )
-    circrna_abundance_ratio = models.IntegerField(
+    circrna_abundance_ratio = models.DecimalField(
+        max_digits=19,
+        decimal_places=6,
         blank=True,
         null=True
     )
